leetcode_43
It is important to know that the (i+j)th position in the product array is determined by ith position in one number and jth position in another. StringBuilder and StringBuffer are both more efficient than String when faced with changing the string continually, while StringBuilder is the most efficient but not thread-safe.